Introducing 3D_wlw2obj.py
This is a non-preservation version of wlw2obj, which creates additional filled 3d meshes in the OBJ files, which can show additional detail not otherwise obvious in the flat OBJ files.

It uses trimesh to handle OBJ exports, and as such, does not preserve the original vertices in their entirety, as the output is optimized. 

The standard wlw2obj.py script handles PRESERVATION of all vertices' data from the original WLW and WLM files.